+++ org.eclipse.tm4e.core/src/main/java/org/eclipse/tm4e/core/internal/theme/css/sac/ISACParserFactory.java
@@ -8,7 +8,7 @@
  * Contributors:
  * Angelo Zerr <angelo.zerr@gmail.com> - initial API and implementation
  *******************************************************************************/
-package org.eclipse.tm4e.core.theme.css;
+package org.eclipse.tm4e.core.internal.theme.css.sac;
 
 import org.eclipse.jdt.annotation.Nullable;
 import org.w3c.css.sac.Parser;
