+++ org.eclipse.tm4e.core/src/main/java/org/eclipse/tm4e/core/theme/css/CSSParser.java
@@ -21,6 +21,7 @@ import org.eclipse.tm4e.core.internal.theme.css.CSSConditionFactory;
 import org.eclipse.tm4e.core.internal.theme.css.CSSDocumentHandler;
 import org.eclipse.tm4e.core.internal.theme.css.CSSSelectorFactory;
 import org.eclipse.tm4e.core.internal.theme.css.ExtendedSelector;
+import org.eclipse.tm4e.core.internal.theme.css.sac.SACParserFactory;
 import org.eclipse.tm4e.core.theme.IStyle;
 import org.w3c.css.sac.CSSException;
 import org.w3c.css.sac.InputSource;
@@ -29,7 +30,7 @@ import org.w3c.css.sac.Selector;
 import org.w3c.css.sac.SelectorList;
 
 /**
- * CSS Parser to parse style for TextMate syntax coloration.
+ * CSS Parser to parse style for TextMate syntax coloring.
  *
  */
 public class CSSParser {
@@ -47,7 +48,7 @@ public class CSSParser {
 	}
 
 	public CSSParser(final InputSource source) throws Exception {
-		this(source, SACParserFactory.newInstance().makeParser());
+		this(source, new SACParserFactory().makeParser());
 	}
 
 	public CSSParser(final String source) throws Exception {
